tenant/slice-activator/requirements.txt
flask
ruamel.yaml
PyYAML
py-json
requests
py-logs
